@@ -26,61 +23,63 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@babel/core": "7.2.2",
+    "@babel/preset-env": "7.2.3",
+    "@babel/preset-react": "7.0.0",
     "assets-webpack-plugin": "3.9.7",
-    "autoprefixer": "9.3.1",
+    "autoprefixer": "9.4.5",
     "axios": "0.18.0",
+    "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.0.1",
-    "babel-loader": "7.1.5",
-    "babel-preset-react": "6.24.1",
-    "clean-webpack-plugin": "1.0.0",
-    "enzyme": "3.7.0",
-    "enzyme-adapter-react-16": "1.7.0",
-    "eslint": "5.9.0",
+    "babel-loader": "8.0.5",
+    "concurrently": "4.1.0",
+    "enzyme": "3.8.0",
+    "enzyme-adapter-react-16": "1.7.1",
+    "eslint": "5.12.1",
     "express": "4.16.4",
     "express-winston": "3.0.1",
-    "extract-text-webpack-plugin": "4.0.0-beta.0",
     "html-webpack-harddisk-plugin": "1.0.1",
     "html-webpack-plugin": "3.2.0",
     "html-webpack-pug-plugin": "0.3.0",
     "isomorphic-fetch": "2.2.1",
-    "node-sass": "4.10.0",
+    "mini-css-extract-plugin": "0.5.0",
+    "node-sass": "4.11.0",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "postcss-loader": "3.0.0",
     "prop-types": "15.6.2",
     "pug": "2.0.3",
-    "pug-loader": "2.4.0",
     "ramda": "0.26.1",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
-    "react-hot-loader": "4.3.12",
-    "react-redux": "5.1.1",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
+    "react-redux": "6.0.0",
     "redux": "4.0.1",
     "redux-thunk": "2.3.0",
     "sass-loader": "7.1.0",
-    "webpack": "4.26.1",
+    "terser-webpack-plugin": "1.2.1",
+    "webpack": "4.28.4",
     "webpack-disk-plugin": "0.0.2",
-    "webpack-merge": "4.1.4",
+    "webpack-merge": "4.2.1",
     "webpack-node-externals": "1.7.2",
     "winston": "3.1.0",
     "write-assets-webpack-plugin": "1.0.4"
   },
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.3",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "babel-preset-env": "1.7.0",
-    "css-loader": "1.0.1",
+    "@babel/node": "7.2.2",
+    "css-loader": "2.1.0",
     "eslint-config-airbnb": "17.1.0",
     "eslint-plugin-babel": "5.3.0",
     "eslint-plugin-import": "2.14.0",
     "eslint-plugin-jsx-a11y": "6.1.2",
-    "eslint-plugin-node": "8.0.0",
+    "eslint-plugin-node": "8.0.1",
     "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-react": "7.11.1",
+    "eslint-plugin-react": "7.12.4",
     "jest": "23.6.0",
-    "nodemon": "1.18.7",
+    "nodemon": "1.18.9",
     "style-loader": "0.23.1",
-    "supertest": "3.3.0",
+    "supertest": "3.4.1",
     "wait-on": "3.2.0",
-    "webpack-cli": "3.1.2",
-    "webpack-dev-server": "3.1.10"
+    "webpack-bundle-analyzer": "3.0.3",
+    "webpack-cli": "3.2.1",
+    "webpack-dev-server": "3.1.14"
   }
 }
