@@ -1,12 +1,11 @@
 import path from 'path';
 import webpack from 'webpack';
-import CleanWebpackPlugin from 'clean-webpack-plugin';
 import nodeExternals from 'webpack-node-externals';
 
 const fileRoot = process.cwd();
 
 const serverPack = {
-	mode: 'development',
+	mode: 'production',
 	stats: 'minimal',
 	target: 'node',
 	entry: './src/server/index.js',
@@ -26,12 +25,8 @@ const serverPack = {
 		],
 	},
 	plugins: [
-		new CleanWebpackPlugin([path.join(fileRoot, 'dist/server')], {
-			root: path.join(__dirname, '../'),
-			verbose: false,
-		}),
 		new webpack.BannerPlugin({ banner: 'require("source-map-support").install();', raw: true, entryOnly: false }),
 	],
 };
 
-module.exports = serverPack;
+export default serverPack;
