@@ -1,62 +1,67 @@
-import ExtractTextPlugin from 'extract-text-webpack-plugin';
-import CleanWebpackPlugin from 'clean-webpack-plugin';
+import path from 'path';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import HtmlWebpackHarddiskPlugin from 'html-webpack-harddisk-plugin';
-import AssetsPlugin from 'assets-webpack-plugin';
 import HtmlWebpackPugPlugin from 'html-webpack-pug-plugin';
-import path from 'path';
-import 'pug-loader';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 
 const fileRoot = process.cwd();
-const extractSass = new ExtractTextPlugin({
-	filename: '[name].[hash].css',
-});
 
-module.exports = {
+const defaultModules = {
+	rules: [
+		{
+			test: /\.scss$/,
+			use: [
+				MiniCssExtractPlugin.loader,
+				'css-loader',
+				'postcss-loader',
+				'sass-loader',
+			],
+		},
+		{
+			test: /.js$/,
+			loader: 'babel-loader',
+			include: path.join(fileRoot, 'src/app'),
+		},
+	],
+};
+
+const defaultPlugins = [
+	new HtmlWebpackPlugin({
+		filename: 'output.pug',
+		alwaysWriteToDisk: true,
+		template: 'src/server/template/index.pug',
+		inject: 'body',
+	}),
+	new HtmlWebpackPugPlugin(),
+	new HtmlWebpackHarddiskPlugin(),
+];
+
+const defaultOptimizations = {
+	runtimeChunk: 'single',
+	splitChunks: {
+		cacheGroups: {
+			vendors: {
+				test: /[\\/]node_modules[\\/]/,
+				name: 'vendors',
+				chunks: 'all',
+				enforce: true,
+			},
+			styles: {
+				name: 'styles',
+				test: /\.scss$/,
+				chunks: 'all',
+				enforce: true,
+			},
+		},
+	},
+};
+
+export default {
+	cache: true,
 	output: {
-		filename: '[name].[hash].js',
 		path: path.resolve(fileRoot, 'dist/public'),
 	},
-	module: {
-		rules: [
-			{
-				test: /\.scss$/,
-				use: extractSass.extract({
-					use: [{
-						loader: 'css-loader',
-					}, {
-						loader: 'sass-loader',
-					}],
-					// use style-loader in development
-					fallback: 'style-loader',
-				}),
-				include: path.join(fileRoot, 'src'),
-			},
-			{
-				test: /.js$/,
-				loader: 'babel-loader',
-				include: path.join(fileRoot, 'src/app'),
-			},
-		],
-	},
-	plugins: [
-		new AssetsPlugin({
-			filename: 'assets.json',
-			path: path.join(fileRoot, '/dist/public'),
-		}),
-		extractSass,
-
-		new HtmlWebpackPlugin({
-			filename: 'output.pug',
-			alwaysWriteToDisk: true,
-			template: 'src/server/template/index.pug',
-			inject: 'body',
-		}),
-		new HtmlWebpackPugPlugin(),
-		new HtmlWebpackHarddiskPlugin(),
-		new CleanWebpackPlugin([path.join(fileRoot, 'dist/public')], {
-			root: path.join(__dirname, '../'),
-			verbose: true,
-		}),
-	],
+	module: defaultModules,
+	plugins: defaultPlugins,
+	optimization: defaultOptimizations,
 };
